cleanup imports and move getVersionFromReadme to base class
